Introduce concrete provider
This listener provider is a more simple version which allows users to
specify which event types it can provide.

It can be useful in some specific cases, for instance if a listener does
not need the event object passed as a parameter (can happen if the
listener only needs to run at a specific stage during runtime, but does
not need event data).

Fixes #10